integrate ssr with firebase functions
