+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "vue-webpack-d244f"
+  }
+}
