+++ firebase.json
@@ -1,0 +1,24 @@
+{
+  "hosting": {
+    "public": "public",
+    "ignore": [
+      "firebase.json",
+      "**/.*",
+      "**/node_modules/**"
+    ],
+    "rewrites": [
+      { "source": "**", "function": "ssr" }
+    ]
+  },
+  "functions": {
+    "source": ".",
+    "ignore": [
+      "components",
+      "static",
+      "store",
+      "style",
+      "**/node_modules/**",
+      "**/.*"
+    ]
+  }
+}
