--- index.pug
@@ -1,18 +1,0 @@
-doctype html
-html(lang="en")
-    
-  head
-    meta(charset="utf-8")
-    meta(http-equiv="X-UA-Compatible", content="IE=edge")
-    meta(name='viewport', content="width=device-width, initial-scale=1")
-    link(rel="apple-touch-icon" sizes="180x180" href="/static/apple-touch-icon.png")
-    link(rel="icon" type="image/png" sizes="32x32" href="/static/favicon-32x32.png")
-    link(rel="icon" type="image/png" sizes="16x16" href="/static/favicon-16x16.png")
-    link(rel="manifest" href="/static/manifest.json")
-    link(rel="mask-icon" href="/static/safari-pinned-tab.svg" color="#5bbad5")
-    meta(name="theme-color" content="#ffffff")
-          
-  body
-    <!--vue-ssr-outlet-->
-    
-
