+++ package.json
@@ -15,18 +15,17 @@
     "build:client": "NODE_ENV='production' ./node_modules/.bin/webpack --config ./build/webpack.client.config.js --progress --hide-modules",
     "build:server": "NODE_ENV='production' ./node_modules/.bin/webpack --config ./build/webpack.server.config.js --progress --hide-modules",
     "build:prerender": "NODE_ENV='production' node build/prerender.js",
-    "build:storybook": "build-storybook --output-dir ./dist/storybook",
-    "build": "rimraf dist && npm run build:client && npm run build:server && npm run build:prerender && npm run build:storybook"
+    "build:storybook": "build-storybook --output-dir ./functions/public/storybook",
+    "build": "rimraf dist && npm run build:client && npm run build:server && npm run build:prerender"
   },
   "devDependencies": {
-    "@storybook/vue": "^3.2.3",
-    "autoprefixer": "^7.1.2",
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
+    "@storybook/vue": "^3.2.11",
+    "autoprefixer": "^7.1.4",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
     "babel-preset-es2015": "^6.24.1",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.4",
-    "express": "^4.15.4",
+    "copy-webpack-plugin": "^4.1.0",
+    "css-loader": "^0.28.7",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.11.2",
     "fontello": "^0.5.0",
@@ -34,31 +33,33 @@
     "memory-fs": "^0.4.1",
     "node-sass": "^4.5.3",
     "postcss-loader": "^2.0.6",
-    "pug": "^2.0.0-rc.3",
-    "rimraf": "^2.6.1",
+    "rimraf": "^2.6.2",
     "sass-loader": "^6.0.6",
-    "sass-resources-loader": "^1.3.0",
+    "sass-resources-loader": "^1.3.1",
     "serviceworker-webpack-plugin": "^0.2.3",
     "style-loader": "^0.18.2",
     "url-loader": "^0.5.9",
-    "vue-loader": "^13.0.4",
-    "vue-server-renderer": "^2.4.2",
-    "vue-template-compiler": "^2.4.2",
-    "webpack": "^3.4.1",
+    "vue-loader": "^13.0.5",
+    "vue-template-compiler": "^2.4.4",
+    "webpack": "^3.6.0",
     "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.18.2",
+    "webpack-hot-middleware": "^2.19.1",
     "webpack-merge": "^4.1.0",
     "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
     "axios": "^0.16.2",
     "bluebird": "^3.5.0",
-    "bulma": "^0.5.0",
+    "bulma": "^0.5.3",
+    "express": "^4.16.1",
+    "firebase-admin": "^5.2.1",
+    "firebase-functions": "^0.6.3",
     "lodash": "^4.17.4",
-    "vue": "^2.4.2",
-    "vue-meta": "^1.0.5",
+    "vue": "^2.4.4",
+    "vue-meta": "^1.2.0",
     "vue-router": "^2.7.0",
-    "vuex": "^2.3.1"
+    "vue-server-renderer": "^2.4.4",
+    "vuex": "^2.4.1"
   },
   "author": "ryantsangai@gmail.com",
   "license": "ISC"
